RaspberryPi_JetsonNano/python/requirements.txt
RPi.GPIO
spidev
numpy
Pillow

icalevents
